eloader: never consider 6A/1phase to be a manual override.
This change for 6A/1 phase is needed as a heurisitc for
when the car is plugged in during the night. Then the charger
will immediately start charging with the last setting (6A/1ph)
and we won't detect this because the current_ampere will be "None"
during the night. This means that one cannot manually override
with 6A/1ph but that is a bit of a pointless setting for a manual
override anyway.

So never consider 6A/1ph as a "manual" setting by the user.